pyproject.toml
[project]
name = "VirtualRecursiveNetwork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adam-atan2>=0.0.3",
    "argdantic>=1.3.3",
    "coolname>=2.2.0",
    "einops>=0.8.1",
    "huggingface-hub>=1.1.7",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "muon-optimizer",
    "ninja>=1.13.0",
    "numba>=0.62.1",
    "omegaconf>=2.3.0",
    "packaging>=25.0",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "setuptools>=80.9.0",
    "setuptools-scm>=9.2.2",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "triton>=3.5.1",
    "wandb>=0.23.0",
    "wheel>=0.45.1",
]

[tool.uv.sources]
muon-optimizer = { git = "https://github.com/KellerJordan/Muon" }
